Fixed FillNA bug
